charm_tests/glance: migrate Amulet tests to Zaza
This patch is the peer change for:

    https://review.opendev.org/#/c/679599/

charm-glance currently has following Amulet tests:

    https://github.com/openstack/charm-glance/blob/master/tests/basic_deployment.py

The 1xx, 2xx and 5xx ones are mostly covered in zaza keystone tests, or could
be skipped in zaza.

Add the only missing 900 one, to test glance restart on config change.

Test option `debug` instead of `use-syslog` to avoid conversion between
`-` and `_`, which is out the scope of this test.

Signed-off-by: Joe Guo <guoqiao@gmail.com>