vault: Await execution prior to validation step
Make it possible for consumers of the ``auto_initialize_no_validation``
function to execute subsequent setup and test code that require vault
and the consumers of the ``certificates`` relation to be ready.

With the current order of execution, it is not possible to use the
configure function in models where ``keystone`` application is not
present.