Refactor swift tests and add storage
The SwiftImageCreateTest was running a pause/resume test for
swift proxy which does not make sense. So, break the proxy specific
tests into their own class and add a class for storage. A
subsequent change to the swift-proxy charm will be needed to add
the new proxy test class to the tests.yaml.