Add support for tests defining resource_cleanup
Add support for tests having a resource_cleanup method. Whether this
is run or not is dictated by the class variable
'run_resource_cleanup'. By default this will be True but any test
can switch it to False so that resources created by the tests are
preserved. Helpful for debugging failures.

FWIW future work might include automatically disabling resource
cleanup if a test fail is detected.