Fix the vault pause/resume test
The pause/resume test paused the lead unit, but then checked
the first unit.  In an HA scenario, the lead unit may not be the
first unit.  This PR changes the test to check the lead unit after
the pause/resume test.

Depends On: https://github.com/openstack-charmers/zaza/pull/327