Test RabbitMQ cluster reboot
Add a test for RabbitMQ cluster reboots. This tests shuts down the
cluster in a controlled order forcing the last broker to become the
primary broker. Then the test starts the secondary brokers asserting
that their units end up in the `blocked` state. Finally, the primary
broker is started restoring the cluster.

Signed-off-by: Nicolas Bock <nicolas.bock@canonical.com>