Store local overcloud CACERT copy in relative path
At present the overcloud CACERT is copied to /tmp and as such it
is not possbile to run multiple tests at once without them
stepping on each other.

Store the copy in a path relative to where the test is executed,
in line with how the SSH keys are stored etc.

Fixes #331