Ensure that the tests directory exists for the local cacert (#373)
A recent change to allow multiple zazas to run at the same time fixed
the cacert file into a local 'tests/' directory.  Unfortunately, that
doesn't exist for every environment where zaza runs (e.g. mojo) and so
this patch ensures that the path exists prior to trying to download into
it.