nova: Conditional security checklist based on presence of vault
The current test expects Nova to never have TLS connections, let's
expect them to be there whenever vault is present.

Remove the 'is-volume-encryption-enabled' assertion as it is not
a property of the Nova security checks. This was previously masked
by the fact that action would always fail due to TLS tests not
being enabled for any bundles.