MySQL Partition Test class
Currently the new partition test_110 is breaking on master due to the
test landing before the charm PR which implemented the change [0] has
landed.

Move the partition test to its own class which can be added as a target
in the charm's tests.yaml. This will not affect master and can be
trivially updated in [0].

[0] https://review.opendev.org/c/openstack/charm-mysql-innodb-cluster/+/779427