Fix output check logic for sync az test
The validation logic for the sync az test checks the specific output of
the command text, but this is dependent on ordering. The code also
validates that the hypervisors are added to the appropriate host
aggregates, so this changes the test to verify the action was successful
and has some output.

Signed-off-by: Billy Olsen <billy.olsen@gmail.com>