Use the availability-zone based on the configured settings
In the attempt to sync the az test, the availability zone is assumed
to be the JUJU_AVAILABILITY_ZONE environment setting. In various
scenarios, this is not what is used for the availability zone of a
compute node.

This change mimics the logic that's contained within the nova-compute
charm to determine which availability zone should be used for the
compute node.

Signed-off-by: Billy Olsen <billy.olsen@gmail.com>