Update nova tests (#563)
Add test for the new nova_cc charm action syncing AZs

The new action is meant to sync the availability zones from the
underlying Juju provider with the OpenStack availability zones.

It is useful within MAAS environments to map the MAAZ AZs to
OpenStack AZs.

Also fix output check logic for sync az test.
The validation logic for the sync az test checks the specific output of
the command text, but this is dependent on ordering. The code also
validates that the hypervisors are added to the appropriate host
aggregates, so this changes the test to verify the action was successful
and has some output.

Also use the availability-zone based on the configured settings.
In the attempt to sync the az test, the availability zone is assumed
to be the JUJU_AVAILABILITY_ZONE environment setting. In various
scenarios, this is not what is used for the availability zone of a
compute node.
This change mimics the logic that's contained within the nova-compute
charm to determine which availability zone should be used for the
compute node.

Signed-off-by: Billy Olsen <billy.olsen@gmail.com>
Co-authored-by: Ionut Balutoiu <ibalutoiu@cloudbasesolutions.com>