Update zaza/__init__.py from zaza
The zaza and zaza-openstack-tests project are currently installed
within the same package.  A side effect of that is that the
zaza/__init__.py from zaza is overwritten by the one in
zaza-openstack-tests.

We ought to find a better way of doing this but we're currently
blocked on using zaza + zaza-openstack-tests on systems with
Python 3.10 installed, and updating this file unblocks us.

Update zaza/__init__.py from openstack-charmers/zaza@c48c955ef7dd8393b54ac96bf72689f757df1364