ovn: Fix ``wrong_bridge_config`` test
The test blindly reverts the ``bridge-interface-mappings``
configuration to an empty string, which would be the wrong thing
to do if it was previously set up by a test configure step.

The test also does not properly populate target_deploy_status and
the test will always wait for the next update status to run,
which may take several minutes, before completing.