Change the ghost share path
Trilio 4.2 generates a hash for the nfs share bases on the path
alone. Unfortunatly the existing Trilio tests set the nfs share
and the ghost share to have the smape path but different source
IPs. For previouse Trilio releases this was fine as the share
hash was different but now they clash and the test fails because it
thinks the share is already mounted.