Don't test nrpe bits on rabbitmq-server. (#793)
When the nrpe application is not deployed in the model skip the test
that checks for the correct configuration of the nagios checks, this is
because at the moment nrpe has no support for jammy, hence is not being
deployed in the jammy-yoga bundles.

Jammy support for nrpe is tracked at http://pad.lv/1968008

Also patch tox.ini and fix a test to allow the tests to pass.