Fix VM retryer
I recently added an option to have zaza delete and relaunch a
guest if it failed to be provisioned or failed a connectivity
test. However, in the code to remove the an instance I
accidentaly called the method that checks that a resource has
gone and not the method that actually does the removal. This
patch fixes that. NOTE: Only the Trilio tests use this
retryer function atm