Respect model in get_swift_storage_topology
Change 0af4c93d causes the target model to be ignored when a
model_name was specified. This in turn is causing the swift
cmr tests to fail as an ip of None is being recorded for
units in a different model.

(cherry picked from commit 17234c9e153b2a23f491f06ab1b37bb174bf9804)