Move add-disk/remove-disk tests to a different class
We are seeing intermittent failures with disk removal, and the most
likely cause for this is that the cluster we are using is too small.
However, we can't simply increase the number of units in the bundles
since there are clients that also run these tests, and having them
all update their bundles can be cumbersome. On the other hand, there's
no reason for non-ceph charms to actually run these tests, so we are
moving them to a separate class, where we can try out the bigger
cluster solution without bothering anyone else.