Switch from json.loads to yaml.safe_load for decoded usernames
In the rotate password test, the results from the action return a list
of strings that are deliminted by single quotes.  This isn't compatible
with json.loads(), but yaml.safe_load() is able to load the string as an
array of strings.

(cherry picked from commit 0fe8e9d666dc79c99217ea86f74e7251474e6c52)