Block until volume reaches 'available' status after detaching.
Before this change during the tearDown process there is an assumption
that a volume can be deleted right after has ben detached, although
there is a transitional state called "detaching" which is more
noticeable on slow environments.

This change makes sure to wait for the volume to reach to the
'available' status before attempting the deletion.

Closes #1048