Add ObjectRetrierWraps for cinder/nova clients in cinder tests
Due to how apache2 & haproxy interact on the cinder unit, this adds the
ObjectRetrier to the cinder and nova clients to allow for the time delay
between the apache2 service starting and the haproxy service noticing
and opening the port.  This allows the test_100... cinder test to
function as (technically) there isn't an error.