Add constraints file(s) for testing environments.
This change introduces a place to keep the constraints needed to build
testing environments. It will be global to all charms under the Charmed
OpenStack umbrella, they will be enforced by zosci at the gate.

These constraints are not meant to be used to build charms, for that the
tooling that should be used is pip-compile.