ovn: Change metric that verifies OVN exporter
It is possible for OVN exporter to run even if it
can't connect to OVN sockets. In that case it will
report only reduced number of metrics. We need to
use metric that confirms that the exporter is successfully
connected to the OVN.

(cherry picked from commit b5ac43acebd2bd1f1512c1677213711ba363307e)
Signed-off-by: Martin Kalcok <martin.kalcok@canonical.com>